pyproject.toml
[tool.poetry]
name = "eage-hackathon"
version = "0.1.0"
description = "project for EAGE hackathon"
authors = ["Mahad Nadeem Janjua"]
readme = "README.md"
packages = [{include = "eage_hackathon"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.23.1"
pandas = "^2.0.2"
numpy = "^1.24.3"
python-pptx = "^0.6.21"
python-docx = "^0.8.11"
pypdf2 = "^3.0.1"
tesseract = "^0.1.3"
streamlit-toggle = "^0.1.3"
pillow = "^9.5.0"
pybase64 = "^1.2.3"
htbuilder = "^0.6.1"
nest-asyncio = "^1.5.6"
sketch = "^0.4.2"
asyncio = "^3.4.3"
requests = "^2.31.0"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
openpyxl = "^3.1.2"
nltk = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
